pyproject.toml
[project]
name = "gppo"
version = "0.1.0"
description = ""
authors = [{name = "Matthijs", email = "matthijs.vanderlende@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium[all]>=1.1.1",
    "matplotlib>=3.10.1",
    "wandb>=0.19.9",
    "hydra-core>=1.3.2",
    "scipy>=1.15.2",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
    "loguru>=0.7.3",
    "notebook>=7.4.0",
    "seaborn>=0.13.2",
    "scikit-learn>=1.6.1",
    "scikit-image>=0.25.2",
    "torch>=2.6.0",
    "ax-platform>=0.5.0",
    "botorch>=0.13.0",
    "gpytorch>=1.14",
    "stable-baselines3>=2.7.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
